mlb-voice-backend/requirements.txt
google-cloud-aiplatform>=1.38.1
google-generativeai>=0.8.4
python-dotenv>=1.0.1
Flask>=3.0.2
flask-cors>=4.0.0
websockets>=12.0
numpy>=1.24.0
sounddevice>=0.4.6
asyncio>=3.4.3
